Implement second puzzle of day 18
It's starting to break down... This one gave the wrong answer, but only off by 1 cube apparently. I missed something but I don't know what. The algorithm seems fine to me and gives the right answer on the example. I just need to add 6 (1 cube) to the answer to get it right. Oh well...