Implement theoretical part 1 of day 24
However it doesn't work. It seems that the precision of my intersection is an issue. I may have to go with floats or a different implementation.
This one came from a different project where the imperative was on performance, but it had a much lower range of inputs than i128.